move test ts
