+++ js/typescript-library/src/__tests__/generate-json-query-merge.ts
@@ -1,10 +1,10 @@
 import * as fs from "fs"
-import {TableField, tableFields} from "./field"
-import {avg, max, sum} from "./measure"
-import {JoinType} from "./query"
-import {from} from "./queryBuilder"
-import {Action, QueryCacheParameter} from "./parameter"
-import {QueryMerge} from "./queryMerge"
+import {TableField, tableFields} from "../field"
+import {avg, max, sum} from "../measure"
+import {JoinType} from "../query"
+import {from} from "../queryBuilder"
+import {Action, QueryCacheParameter} from "../parameter"
+import {QueryMerge} from "../queryMerge"
 
 export function generateFromQueryMerge() {
   const fields = tableFields(["a", "b"])
@@ -22,5 +22,5 @@ export function generateFromQueryMerge() {
 
   const q = new QueryMerge(query1).join(query2, JoinType.LEFT).join(query3, JoinType.INNER).withParameter(new QueryCacheParameter(Action.NOT_USE))
   const data = JSON.stringify(q)
-  fs.writeFileSync('json/build-from-query-merge.json', data)
+  fs.writeFileSync('./json/build-from-query-merge.json', data)
 }
