+++ js/typescript-library/src/main.ts
@@ -1,10 +1,10 @@
-import {generateFromQueryDto} from "./generate-json-queryDto"
-import {generateJsonQuery} from "./generate-json-query"
-import {generateFromQueryMerge} from "./generate-json-query-merge"
-import {generateFromQueryPivot} from "./generate-json-query-pivot"
-import {generateFromQueryMergePivot} from "./generate-json-query-merge-pivot"
-import {generateFromQueryJoin} from "./generate-json-query-join"
-import {generateQueryResults} from "./generate-json-query-results"
+import {generateFromQueryDto} from "./__tests__/generate-json-queryDto"
+import {generateJsonQuery} from "./__tests__/generate-json-query"
+import {generateFromQueryMerge} from "./__tests__/generate-json-query-merge"
+import {generateFromQueryPivot} from "./__tests__/generate-json-query-pivot"
+import {generateFromQueryMergePivot} from "./__tests__/generate-json-query-merge-pivot"
+import {generateFromQueryJoin} from "./__tests__/generate-json-query-join"
+import {generateQueryResults} from "./__tests__/generate-json-query-results"
 
 generateFromQueryDto()
 generateJsonQuery()
